+++ example-adaptive/build.gradle
@@ -1,5 +1,5 @@
 apply plugin: 'com.android.application'
-apply plugin: 'com.github.gfx.ribbonizer'
+apply plugin: 'com.shogo82148.ribbonizer'
 
 android {
     compileSdkVersion 28
