+++ plugin/src/main/groovy/com/shogo82148/ribbonizer/CustomColorRibbonBuilder.java
@@ -1,11 +1,10 @@
-package com.github.gfx.ribbonizer;
+package com.shogo82148.ribbonizer;
 
 import com.android.build.gradle.api.ApplicationVariant;
-import com.github.gfx.ribbonizer.filter.ColorRibbonFilter;
-import com.github.gfx.ribbonizer.resource.Resource;
+import com.shogo82148.ribbonizer.filter.ColorRibbonFilter;
+import com.shogo82148.ribbonizer.resource.Resource;
 
 import java.awt.Color;
-import java.awt.image.BufferedImage;
 import java.io.File;
 import java.util.function.Consumer;
