+++ plugin/src/main/groovy/com/shogo82148/ribbonizer/FilterBuilder.java
@@ -1,7 +1,7 @@
-package com.github.gfx.ribbonizer;
+package com.shogo82148.ribbonizer;
 
 import com.android.build.gradle.api.ApplicationVariant;
-import com.github.gfx.ribbonizer.resource.Resource;
+import com.shogo82148.ribbonizer.resource.Resource;
 
 import java.io.File;
 import java.util.function.BiFunction;
