+++ plugin/src/main/groovy/com/shogo82148/ribbonizer/YellowRibbonBuilder.java
@@ -1,17 +1,16 @@
-package com.github.gfx.ribbonizer;
+package com.shogo82148.ribbonizer;
 
 import com.android.build.gradle.api.ApplicationVariant;
-import com.github.gfx.ribbonizer.filter.ColorRibbonFilter;
-import com.github.gfx.ribbonizer.resource.Resource;
+import com.shogo82148.ribbonizer.filter.ColorRibbonFilter;
+import com.shogo82148.ribbonizer.resource.Resource;
 
 import java.awt.*;
-import java.awt.image.BufferedImage;
 import java.io.File;
 import java.util.function.Consumer;
 
-public class GreenRibbonBuilder implements FilterBuilder {
+public class YellowRibbonBuilder implements FilterBuilder {
 
-    private static Color COLOR = new Color(0, 0x72, 0, 0x99);
+    private static final Color COLOR = new Color(0xff, 0x76, 0, 0x99);
 
     @Override
     public Consumer<Resource> apply(ApplicationVariant variant, File iconFile) {
