+++ plugin/src/main/groovy/com/shogo82148/ribbonizer/filter/GrayScaleFilter.java
@@ -1,9 +1,9 @@
-package com.github.gfx.ribbonizer.filter;
+package com.shogo82148.ribbonizer.filter;
 
-import com.github.gfx.ribbonizer.resource.Filter;
-import com.github.gfx.ribbonizer.resource.ImageAdaptiveIcon;
-import com.github.gfx.ribbonizer.resource.ImageIcon;
-import com.github.gfx.ribbonizer.resource.Resource;
+import com.shogo82148.ribbonizer.resource.Filter;
+import com.shogo82148.ribbonizer.resource.ImageAdaptiveIcon;
+import com.shogo82148.ribbonizer.resource.ImageIcon;
+import com.shogo82148.ribbonizer.resource.Resource;
 
 import java.awt.image.BufferedImage;
 import java.util.function.Consumer;
