+++ plugin/src/main/groovy/com/shogo82148/ribbonizer/resource/ImageAdaptiveIcon.java
@@ -1,4 +1,4 @@
-package com.github.gfx.ribbonizer.resource;
+package com.shogo82148.ribbonizer.resource;
 
 import java.awt.image.BufferedImage;
 import java.io.File;
