+++ plugin/src/main/groovy/com/shogo82148/ribbonizer/resource/Resource.java
@@ -1,4 +1,4 @@
-package com.github.gfx.ribbonizer.resource;
+package com.shogo82148.ribbonizer.resource;
 
 import java.io.File;
 import java.io.IOException;
