--- plugin/src/main/resources/META-INF/gradle-plugins/com.github.gfx.ribbonizer.properties
@@ -1,1 +1,0 @@
-implementation-class=com.github.gfx.ribbonizer.plugin.RibbonizerPlugin
