+++ plugin/src/main/resources/META-INF/gradle-plugins/com.shogo82148.ribbonizer.properties
@@ -1,0 +1,1 @@
+implementation-class=com.shogo82148.ribbonizer.plugin.RibbonizerPlugin
